fix: refactor to better encapsulate switchenv component
